backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
flask-sqlalchemy
python-dotenv
groq
openai
pydub
numpy
webrtcvad

